Fix: Convert activations to float32 in _collect_activations
- Add .float() when extracting last token activations
- This ensures data is float32 before BCE loss training
- Fixes CUDA assertion failure at layer 24 training
- Complements existing fixes in training/validation loops

+++ mira/attack/ssr/probe_ssr.py
@@ -410,8 +410,8 @@ class ProbeSSR(SSRAttack):
                     else:
                         continue
                     
-                    # Use last token
-                    last_token_act = act[0, -1, :].cpu()  # [d_model]
+                    # Use last token - convert to float32 to prevent BCE issues with float16
+                    last_token_act = act[0, -1, :].cpu().float()  # [d_model]
                     layer_acts[layer_idx].append(last_token_act)
         
         finally:
